@@ -5,11 +5,17 @@
   "dependencies": {
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
-    "@mui/material": "^6.3.0",
+    "@mui/material": "^6.3.1",
     "@mui/x-date-pickers": "^7.23.3",
+    "axios": "^1.7.9",
+    "chart.js": "^4.4.7",
     "cra-template": "1.2.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "html-react-parser": "^5.2.2",
+    "react": "^18.3.1",
+    "react-alice-carousel": "^2.9.1",
+    "react-chartjs-2": "^5.3.0",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^7.1.1",
     "react-scripts": "^5.0.1",
     "web-vitals": "^4.2.4"
   },
