@@ -1,12 +1,16 @@
 import React from 'react';
 import Banner from '../components/Banner/Banner';
+import CoinsTable from "../components/CoinsTable";
 
 const Homepage = () => {
   return (
-    <div>
-      <Banner></Banner>
-    </div>
-  )
-}
+    <>
+    <Banner />
+    <CoinsTable />
+  </>
+);
+};
+
+    
 
 export default Homepage;
