@@ -1,66 +1,64 @@
-import { Container, makeStyles, Typography } from '@mui/material';
-import React from 'react'
+import { Container, Typography, Box } from '@mui/material';
+import React from 'react';
+import Carousel from "./Carousel";
+import { styled } from '@mui/system';
 
-const useStyles=makeStyles((theme) => ({
-  banner: {
-    backgroundImage: "url(./banner3.jpg)",
-  },
-  bannerContent: {
-    height: 400,
-    display: "flex",
-    flexDirection: "column",
-    paddingTop: 25,
-    justifyContent: "space-around",
+// Styled components
+const BannerWrapper = styled('div')({
+  backgroundImage: "url(./banner3.jpg)",
+});
 
-  },
-  tagline: {
-display: "flex",
-flexDirection: "column",
-justifyContent: "center",
-height: "40%",
-textAlign: "center",
-  },
-  carousel: {
-    height: "50%",
-    display: "flex",
-    alignItems: "center",
-  },
-}));
+const BannerContent = styled(Container)({
+  height: 400,
+  display: "flex",
+  flexDirection: "column",
+  paddingTop: 25,
+  justifyContent: "space-around",
+});
+
+const Tagline = styled('div')({
+  display: "flex",
+  flexDirection: "column",
+  justifyContent: "center",
+  height: "40%",
+  textAlign: "center",
+});
+
+const BannerTitle = styled(Typography)({
+  fontWeight: "bold",
+  marginBottom: 15,
+  fontFamily: "Montserrat",
+});
+
+const BannerSubtitle = styled(Typography)({
+  color: "darkgrey",
+  textTransform: "capitalize",
+  fontFamily: "Montserrat",
+});
+
+const CarouselContainer = styled('div')({
+  height: "50%",
+  display: "flex",
+  alignItems: "center",
+});
 
 function Banner() {
-  const classes = useStyles();
   return (
-    <div className={classes.banner}>
-      <Container className={classes.bannerContent}>
-        <div className={classes.tagline}>
-          <Typography
-          variant="h2"
-          style={{
-            fontWeight: "bold",
-            marginBottom: 15,
-            fontFamily: "Montserrat",
-
-          }}
-          >
+    <BannerWrapper>
+      <BannerContent>
+        <Tagline>
+          <BannerTitle variant="h2">
             Crypto Trader
-          </Typography>
-          <Typography
-          variant="subtitle2"
-          style={{
-            color: "darkgrey",
-            textTransform: "capitalize",
-            fontFamily: "Montserrat",
-          }}
-          >
+          </BannerTitle>
+          <BannerSubtitle variant="subtitle2">
             scale through and know more about your crypto coins
-
-          </Typography>
-
-        </div>
-        <Carousel></Carousel>
-
-      </Container>
-    </div>
+          </BannerSubtitle>
+        </Tagline>
+        <CarouselContainer>
+          <Carousel />
+        </CarouselContainer>
+      </BannerContent>
+    </BannerWrapper>
   );
 }
 
